update to harmony version with support for metadata, update associated code to pull in metadata

+++ backend/FwLite/MiniLcm/Media/FileMetadata.cs
@@ -5,4 +5,4 @@ public record LcmFileMetadata(
     string MimeType,
     string? Author = null,
     DateTimeOffset? UploadDate = null,
-    int? SizeInBytes = null);
+    long? SizeInBytes = null);
